# submission/openapi.yaml
openapi: "3.0.3"
info:
  title: HR Management System API
  version: "2.0.0"
  description: |
    REST API for the Stella Technology HR Management System.

    ## Roles
    | Role | Access |
    |------|--------|
    | **ADMIN** | Registers employees & departments, views all leaves and payroll (read-only on those), approves HR employees' leaves |
    | **HR** | Manages their department only — approves employee leaves, generates payroll; submits own leaves (approved by Admin) |
    | **EMPLOYEE** | Applies for leave, views own leaves, payroll, and profile |

    ## Authentication
    1. Call `POST /api/auth/login` with your credentials
    2. Copy the `token` from the response
    3. Click **Authorize** → enter `Bearer <token>`
    4. All subsequent requests include the token automatically

  contact:
    name: Hassan Raheem
    email: hmarwat49@gmail.com
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EmployeeRequest:
      type: object
      required: [name, email, phone, gender, departmentId, designation, salary]
      properties:
        name:         { type: string,  example: "Jane Smith" }
        email:        { type: string,  format: email, example: "jane.smith@company.com" }
        phone:        { type: string,  example: "+92-301-0000001", minLength: 10, maxLength: 15 }
        gender:       { type: string,  example: "FEMALE" }
        dateOfBirth:  { type: string,  format: date, example: "1998-03-12" }
        address:      { type: string,  example: "456 Park Avenue, Lahore" }
        departmentId: { type: string,  example: "DEPT-0001" }
        designation:  { type: string,  example: "Software Engineer" }
        dateOfJoining: { type: string,  format: date, example: "2026-09-01" }
        salary:       { type: number,  example: 70000 }
        allowances:   { type: number,  example: 5000,  default: 0 }
        deductions:   { type: number,  example: 3000,  default: 0 }
        employeeType: { type: string,  enum: [EMPLOYEE, MANAGER, HR] }
        userRole:     { type: string,  enum: [ADMIN, HR, EMPLOYEE] }
    EmployeeUpdateRequest:
      type: object
      description: All fields are optional — only include the fields you want to change.
      properties:
        name:         { type: string }
        email:        { type: string, format: email }
        phone:        { type: string }
        gender:       { type: string }
        dateOfBirth:  { type: string, format: date }
        address:      { type: string }
        departmentId: { type: string }
        designation:  { type: string }
        dateOfJoining: { type: string, format: date }
        salary:       { type: number }
        allowances:   { type: number }
        deductions:   { type: number }
        employeeType: { type: string, enum: [EMPLOYEE, MANAGER, HR] }
        userRole:     { type: string, enum: [ADMIN, HR, EMPLOYEE] }
    EmployeeResponse:
      type: object
      properties:
        employeeId:   { type: string,  example: "EMP-0001" }
        name:         { type: string,  example: "Jane Smith" }
        email:        { type: string,  example: "jane.smith@company.com" }
        phone:        { type: string }
        gender:       { type: string }
        dateOfBirth:  { type: string,  format: date }
        address:      { type: string }
        departmentId: { type: string,  example: "DEPT-0001" }
        designation:  { type: string }
        dateOfJoining: { type: string,  format: date }
        salary:       { type: number }
        allowances:   { type: number }
        deductions:   { type: number }
        status:       { type: string,  enum: [ACTIVE, INACTIVE] }
        employeeType: { type: string,  enum: [EMPLOYEE, MANAGER, HR] }
        userRole:     { type: string,  enum: [ADMIN, HR, EMPLOYEE] }
        createdAt:    { type: string,  format: date-time }
        updatedAt:    { type: string,  format: date-time }
    EmployeePage:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/EmployeeResponse' }
        totalElements: { type: integer, example: 42 }
        totalPages:    { type: integer, example: 5 }
        number:        { type: integer, example: 0, description: "Current page (0-indexed)" }
        size:          { type: integer, example: 10 }
    PerformanceNoteRequest:
      type: object
      required: [note]
      properties:
        note: { type: string, example: "Excellent performance in Q3 2026" }
    EmployeeHistoryResponse:
      type: object
      properties:
        historyId:     { type: string,  example: "HIST-0001" }
        employeeId:    { type: string,  example: "EMP-0001" }
        type:          { type: string,  enum: [PROMOTION, DESIGNATION_CHANGE, PERFORMANCE_NOTE] }
        previousValue: { type: string,  nullable: true, example: "EMPLOYEE" }
        newValue:      { type: string,  nullable: true, example: "MANAGER" }
        note:          { type: string,  nullable: true, description: "Populated only for PERFORMANCE_NOTE" }
        recordedBy:    { type: string,  example: "EMP-0001", description: "employeeId of the admin who made the change" }
        recordedAt:    { type: string,  format: date-time }

    # ── Department ────────────────────────────────────────────────────────
...
    LeaveResponse:
      type: object
      properties:
        leaveId:    { type: string,  example: "LVE-0001" }
        employeeId: { type: string,  example: "EMP-0002" }
        leaveType:  { type: string,  enum: [ANNUAL, CASUAL, SICK, UNPAID] }
        startDate:  { type: string,  format: date }
        endDate:    { type: string,  format: date }
        totalDays:  { type: integer, example: 5 }
        reason:     { type: string,  nullable: true }
        status:     { type: string,  enum: [PENDING, APPROVED, REJECTED] }
        appliedAt:  { type: string,  format: date-time }
        updatedAt:  { type: string,  format: date-time }
...
    PayrollResponse:
      type: object
      properties:
        payrollId:     { type: string,  example: "PAY-0001" }
        employeeId:    { type: string,  example: "EMP-0002" }
        month:         { type: integer, example: 9 }
        year:          { type: integer, example: 2026 }
        basicSalary:   { type: number,  example: 70000 }
        allowances:    { type: number,  example: 5000 }
        deductions:    { type: number,  example: 3000 }
        netSalary:     { type: number,  example: 72000, description: "basicSalary + allowances - deductions" }
        paymentStatus: { type: string,  enum: [PAID, UNPAID] }
        generatedAt:   { type: string,  format: date-time }
        updatedAt:     { type: string,  format: date-time }

    # ── Errors ────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      properties:
        status:    { type: integer, example: 400 }
        message:   { type: string,  example: "Validation failed" }
        timestamp: { type: string,  format: date-time }
  responses:
    Unauthorized:
      description: "401 — Missing or expired JWT token"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: "403 — Authenticated but role does not have permission"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: "404 — Resource not found"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: "400 — Validation error or business rule violation"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
security:
  - BearerAuth: []
...
paths:
...
  /api/employees/me:
    get:
      tags: [Employees]
      summary: Get own employee profile
      description: Returns the logged-in user's employee profile. Accessible by all roles.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Own employee profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /api/employees:
    get:
      tags: [Employees]
      summary: List employees (paginated)
      description: |
        Returns a paginated list of employees.
        - **ADMIN**: sees all employees. All query params apply.
        - **HR**: automatically scoped to their own department (departmentId param is ignored).
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
          description: Page number (0-indexed)
        - name: size
          in: query
          schema: { type: integer, default: 10 }
          description: Results per page
        - name: name
          in: query
          schema: { type: string }
          description: Filter by name (partial, case-insensitive)
        - name: departmentId
          in: query
          schema: { type: string }
          description: Filter by department ID (ADMIN only — ignored for HR)
        - name: status
          in: query
          schema: { type: string, enum: [ACTIVE, INACTIVE] }
          description: Filter by employment status
        - name: userRole
          in: query
          schema: { type: string, enum: [ADMIN, HR, EMPLOYEE] }
          description: Filter by role
      responses:
        "200":
          description: Paginated employee list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeePage' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Employees]
      summary: Register new employee
      description: |
        Creates both the employee profile and their login user account in one step.
        - HR employees may be registered without a department (they are placed in Unassigned/DEPT-0000 until a department is created for them).
        - Requires **ADMIN** or **HR**.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmployeeRequest' }
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /api/employees/{employeeId}:
    get:
      tags: [Employees]
      summary: Get employee by ID
      description: Returns a single employee. Accessible by ADMIN, HR, and EMPLOYEE.
      security:
        - BearerAuth: []
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string, example: "EMP-0001" }
      responses:
        "200":
          description: Employee found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

    put:
      tags: [Employees]
      summary: Update employee
      description: |
        Partially updates an employee's profile. All fields are optional — send only the fields you want to change.
        Requires **ADMIN** or **HR**.
      security:
        - BearerAuth: []
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string, example: "EMP-0001" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmployeeUpdateRequest' }
            example:
              designation: "Senior Software Engineer"
              salary: 85000
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Employees]
      summary: Delete employee
      description: |
        Deletes the employee along with their user account, leave balance, and all leave requests.
        Requires **ADMIN**.
      security:
        - BearerAuth: []
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string, example: "EMP-0001" }
      responses:
        "204":
          description: Employee deleted successfully (no body)
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /api/employees/{employeeId}/type:
    patch:
      tags: [Employees]
      summary: Change employee type
      description: |
        Changes an employee's type (EMPLOYEE / MANAGER / HR).
        Automatically creates a `PROMOTION` or `DESIGNATION_CHANGE` history record.
        Requires **ADMIN**.
      security:
        - BearerAuth: []
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string, example: "EMP-0001" }
        - name: employeeType
          in: query
          required: true
          schema: { type: string, enum: [EMPLOYEE, MANAGER, HR] }
      responses:
        "200":
          description: Employee type updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /api/employees/{employeeId}/performance-notes:
    post:
      tags: [Employees]
      summary: Add performance note
      description: Adds a free-text performance note to the employee's history. Requires **ADMIN**.
      security:
        - BearerAuth: []
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string, example: "EMP-0001" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PerformanceNoteRequest' }
      responses:
        "201":
          description: Performance note added
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeHistoryResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /api/employees/{employeeId}/history:
    get:
      tags: [Employees]
      summary: Get employee history
      description: |
        Returns the promotion, designation change, and performance note history for an employee.
        Accessible by ADMIN, HR, and the employee themselves.
      security:
        - BearerAuth: []
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string, example: "EMP-0001" }
        - name: type
          in: query
          required: false
          schema: { type: string, enum: [PROMOTION, DESIGNATION_CHANGE, PERFORMANCE_NOTE] }
          description: Filter by history record type
      responses:
        "200":
          description: History records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EmployeeHistoryResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
  /api/employees/{employeeId}/leaves:
    get:
      tags: [Employees]
      summary: Get employee's leave history
      description: Returns all leave requests for the specified employee. Accessible by ADMIN, HR, and the employee themselves.
      security:
        - BearerAuth: []
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string, example: "EMP-0001" }
      responses:
        "200":
          description: Leave history
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LeaveResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
  /api/employees/{employeeId}/payroll:
    get:
      tags: [Employees]
      summary: Get employee's payroll history
      description: Returns all payroll records for the specified employee. Accessible by ADMIN, HR, and the employee themselves.
      security:
        - BearerAuth: []
      parameters:
        - name: employeeId
          in: path
          required: true
          schema: { type: string, example: "EMP-0001" }
      responses:
        "200":
          description: Payroll records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PayrollResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ════════════════════════════════════════════════════════════
  #  DEPARTMENTS
  # ════════════════════════════════════════════════════════════
...